+++ src/Bootstrap.php
@@ -68,6 +68,7 @@ class Bootstrap {
 		add_filter( 'gu_get_language_pack_json', [ $this, 'set_language_pack_json' ], 10, 4 );
 		add_filter( 'gu_post_process_language_pack_package', [ $this, 'process_language_pack_data' ], 10, 4 );
 		add_filter( 'gu_get_git_icon_data', [ $this, 'set_git_icon_data' ], 10, 2 );
+		add_filter( 'gua_addition_types', [ $this, 'add_addition_types' ], 10, 1 );
 	}
 
 	/**
@@ -389,5 +390,16 @@ class Bootstrap {
 		);
 
 		return $icon_data;
+	}
+
+	/**
+	 * Add repository types to Git Updater Additions.
+	 *
+	 * @param array $addition_types Array of Git Updater Additions repository types.
+	 *
+	 * @return array
+	 */
+	public function add_addition_types( $addition_types ) {
+		return array_merge( $addition_types, [ 'gitlab_plugin', 'gitlab_theme' ] );
 	}
 }