+++ CHANGES.md
@@ -1,4 +1,5 @@
 [unreleased]
+* use new filter to add repository types to Git Updater Additions
 
 #### 1.1.1 / 2021-9-24
 * switch from `PaND` to `WP_Dismiss_Notice`